Implemented ProjectionAll classes in Python.
Decided to use PyTorch, pybind11, CUDA to make ProjectionAll.

ProjectionAll will have no dependencies on OpenBCSim.

This is just easier to design for because I need a different set
of features than OpenBCSim is focused on.

Currently, Python classes and CUDA kernel are implemented.

Next up: Implement the pybind11 glue that interfaces the two.